chore: prune cruft + restructure auth services
Multi-part cleanup pass after the auth/GCP work landed. Strict zero
behavior change — only file moves, deletions, and renames.

## Deletions (vestigial / wrong-place files)

- `/main.py` — empty 0-byte file at repo root, unreferenced.
- `/backend/main.py` — leftover FastAPI hello-world stub. Django is the
  entrypoint (`config.wsgi`); nothing imported this.
- `/backend/package-lock.json` — an npm artifact accidentally created in
  the Python project root.
- `/infra/terraform/main.tf` — pre-modular monolithic terraform; replaced
  by the focused files (`apis.tf`, `cloud_run_*.tf`, `secrets.tf`, …).
- `/infra/terraform/*.tfplan` — ephemeral plan artifacts, already in
  `.gitignore`.
- `/docker-compose.yml` and `/.env.docker` — local-only docker-compose
  setup that was never wired in; production uses GCP Cloud Run.
- `/frontend/.env.local` — leftover from a `vercel link` session.
- `/sdks/python/apilenss/` — entire directory of 0-byte stub files.
  The PyPI package name stays `apilenss` (set in `sdks/python/pyproject.toml`
  via `name = "apilenss"`); only the real implementation in
  `sdks/python/apilens/` is bundled into the wheel now.

## Restructure

- `backend/core/auth/authentication.py` → `backend/apps/auth/authentication.py`.
  The file depends on `apps.users.models` and `apps.auth.models`, which
  was a layering violation per CLAUDE.md ("core never imports from apps").
  It belongs in `apps/auth/`, not `core/`. Updated 5 router imports + the
  `core/auth/__init__.py` re-export list. All in-method deferred imports
  are now top-level since the circular-import risk is gone.

- `backend/apps/auth/services.py` (1227 lines, 8 service classes)
  → `backend/apps/auth/services/` package.
  - `_constants.py` — shared timedeltas, rate limits, `hash_token` helper.
  - `tokens.py` — TokenService (refresh tokens, session enumeration).
  - `magic_link.py` — MagicLinkService.
  - `password_reset.py` — PasswordResetService.
  - `auth.py` — AuthService (top-level orchestration).
  - `api_keys.py` — ApiKeyService + API_KEY_PREFIX constant.
  - `passkey.py` — PasskeyService (WebAuthn).
  - `two_factor.py` — TwoFactorService (TOTP + backup codes).
  - `recovery.py` — RecoveryService (48h delayed account recovery).
  - `__init__.py` re-exports every public symbol, so existing
    `from apps.auth.services import TokenService` callers (16 sites)
    keep working unchanged.

- `backend/pyproject.toml` — dropped unused `apilenss` dependency
  (declared but never imported by any backend code).

## Newly tracked

- `sidecar-testing/` — local integration testing apps for the Python SDK
  across FastAPI / Flask / Django-Ninja. Was gitignored. Per-app venvs,
  dotenvs, and sqlite files stay ignored.

- `FOLLOWUPS.md` — items deliberately deferred (settings env-var fallback
  chains, backend test coverage, eslint-config-next compat, Supabase
  pooler switch, `apps/endpoints` migration squash).

## Verification

- `python -m py_compile` clean on every new file.
- Django: `manage.py check --deploy` shows only the usual dev-mode SSL/HSTS
  warnings; no model or import errors.
- `ruff check apps/auth/ core/auth/` — all checks passed.
- `from apps.auth.services import (TokenService, MagicLinkService,
  PasswordResetService, AuthService, ApiKeyService, PasskeyService,
  TwoFactorService, RecoveryService, ACCESS_TOKEN_LIFETIME, ...)`
  — resolves cleanly via the package's `__init__.py`.
