--- sdks/python/apilenss/blacksheep.py
@@ -1,2 +1,0 @@
-from apilens.blacksheep import *  # noqa: F401,F403
-
