+++ sidecar-testing/django-ninja/.env.example
@@ -1,0 +1,8 @@
+APILENS_API_KEY=
+APILENS_PROJECT_SLUG=
+APILENS_APP_ID=
+APILENS_BASE_URL=http://localhost:8000/api/v1
+APILENS_ENVIRONMENT=development
+PORT=8013
+DJANGO_SECRET_KEY=sidecar-dev-key
+DJANGO_DEBUG=true
