+++ sidecar-testing/django-ninja/README.md
@@ -1,0 +1,11 @@
+# Django Ninja Sidecar
+
+Run:
+
+```bash
+cp .env.example .env
+# set APILENS_API_KEY, APILENS_PROJECT_SLUG, and APILENS_APP_ID in .env
+source .venv/bin/activate
+python manage.py migrate
+python manage.py runserver 0.0.0.0:${PORT:-8013}
+```
