+++ sidecar-testing/django-ninja/project/settings.py
@@ -1,0 +1,55 @@
+import os
+from pathlib import Path
+
+from dotenv import load_dotenv
+
+BASE_DIR = Path(__file__).resolve().parent.parent
+load_dotenv(BASE_DIR / ".env")
+
+SECRET_KEY = os.getenv("DJANGO_SECRET_KEY", "sidecar-dev-key")
+DEBUG = os.getenv("DJANGO_DEBUG", "true").lower() == "true"
+ALLOWED_HOSTS = ["*"]
+
+INSTALLED_APPS = [
+    "django.contrib.contenttypes",
+    "django.contrib.auth",
+    "django.contrib.sessions",
+    "django.contrib.messages",
+    "django.contrib.staticfiles",
+]
+
+MIDDLEWARE = [
+    "django.middleware.security.SecurityMiddleware",
+    "django.contrib.sessions.middleware.SessionMiddleware",
+    "django.middleware.common.CommonMiddleware",
+    "django.contrib.auth.middleware.AuthenticationMiddleware",
+    "django.contrib.messages.middleware.MessageMiddleware",
+    "apilens.django.ApiLensDjangoMiddleware",
+]
+
+ROOT_URLCONF = "project.urls"
+
+TEMPLATES = []
+WSGI_APPLICATION = "project.wsgi.application"
+ASGI_APPLICATION = "project.asgi.application"
+
+DATABASES = {
+    "default": {
+        "ENGINE": "django.db.backends.sqlite3",
+        "NAME": BASE_DIR / "db.sqlite3",
+    }
+}
+
+LANGUAGE_CODE = "en-us"
+TIME_ZONE = "UTC"
+USE_I18N = True
+USE_TZ = True
+
+STATIC_URL = "/static/"
+DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
+
+APILENS_API_KEY = os.getenv("APILENS_API_KEY", "")
+APILENS_PROJECT_SLUG = os.getenv("APILENS_PROJECT_SLUG", "")
+APILENS_APP_ID = os.getenv("APILENS_APP_ID", "")
+APILENS_BASE_URL = os.getenv("APILENS_BASE_URL", "https://api.apilens.ai/api/v1")
+APILENS_ENVIRONMENT = os.getenv("APILENS_ENVIRONMENT", "development")
