+++ sidecar-testing/django-ninja/requirements.txt
@@ -1,0 +1,4 @@
+django>=5.0
+django-ninja>=1.1.0
+python-dotenv>=1.0.1
+apilenss[django]>=0.1.5
