+++ sidecar-testing/fastapi/.env.example
@@ -1,0 +1,6 @@
+APILENS_API_KEY=
+APILENS_PROJECT_SLUG=
+APILENS_APP_ID=
+APILENS_BASE_URL=http://localhost:8000/api/v1
+APILENS_ENVIRONMENT=development
+PORT=8011
