+++ sidecar-testing/flask/README.md
@@ -1,0 +1,10 @@
+# Flask Sidecar
+
+Run:
+
+```bash
+cp .env.example .env
+# set APILENS_API_KEY, APILENS_PROJECT_SLUG, and APILENS_APP_ID in .env
+source .venv/bin/activate
+python app/main.py
+```
