+++ sidecar-testing/flask/requirements.txt
@@ -1,0 +1,3 @@
+flask>=3.0.0
+python-dotenv>=1.0.1
+apilenss[flask]>=0.1.5
